Move DataFrame `dispatch.registers` to their own file (#7503)
