+++ dask/dataframe/reshape.py
@@ -2,12 +2,12 @@ import sys
 
 import numpy as np
 import pandas as pd
-from pandas.api.types import is_list_like
+from pandas.api.types import is_list_like, is_scalar
 
 from ..utils import M
 from . import methods
 from .core import DataFrame, Series, apply_concat_apply, map_partitions
-from .utils import has_known_categories, is_scalar
+from .utils import has_known_categories
 
 ###############################################################
 # Dummies
